Moved difficulty calculation into own application
