+++ src/module/combat/combat.js
@@ -1,5 +1,6 @@
 import { DiceSFRPG } from "../dice.js";
 import RollContext from "../rolls/rollcontext.js";
+import { CombatDifficulty } from "../apps/combat-difficulty.js";
 
 /*
 The following hooks were added:
@@ -556,407 +557,6 @@ export class CombatSFRPG extends Combat {
         return this.getIndexOfFirstUndefeatedCombatant() === null;
     }
 
-    getStarshipEncounterInfo() {
-        // Performs starship encounter difficulty calculations and populates flags
-        if (!this.flags.sfrpg) {
-            this.flags.sfrpg = {};
-        }
-
-        const [playerShips, playerShipTiers, enemyShips, enemyShipTiers] = this.parseShips();
-        this.flags.sfrpg.playerShips = playerShips;
-        this.flags.sfrpg.playerShipTiers = playerShipTiers;
-        this.flags.sfrpg.enemyShips = enemyShips;
-        this.flags.sfrpg.enemyShipTiers = enemyShipTiers;
-
-        const [playerTierEffective, playerTierRound, playerTierString] = this.calculatePlayerShipEffectiveTier(playerShipTiers);
-        this.flags.sfrpg.playerTier = {};
-        this.flags.sfrpg.playerTier.effective = playerTierEffective;
-        this.flags.sfrpg.playerTier.round = playerTierRound;
-        this.flags.sfrpg.playerTier.string = playerTierString;
-
-        const [enemyTierEffective, enemyTierRound, enemyTierString] = this.calculateEnemyShipEffectiveTier(enemyShipTiers);
-        this.flags.sfrpg.enemyTier = {};
-        this.flags.sfrpg.enemyTier.effective = enemyTierEffective;
-        this.flags.sfrpg.enemyTier.round = enemyTierRound;
-        this.flags.sfrpg.enemyTier.string = enemyTierString;
-
-        console.log(this);
-        const [difficulty, XPValue, wealth] = this.calculateShipChallenge();
-        this.flags.sfrpg.difficulty = difficulty;
-        this.flags.sfrpg.XPValue = XPValue;
-        this.flags.sfrpg.wealth = wealth;
-    }
-
-    parseShips() {
-        let playerShips = [];
-        let playerShipTiers = [];
-        let enemyShips = [];
-        let enemyShipTiers = [];
-
-        // split combatants into allies and enemies
-        for (const combatant of this.combatants) {
-            if (combatant.actor.type === "starship") {
-                if (combatant.hasPlayerOwner) {
-                    playerShips.push(combatant);
-                    playerShipTiers.push(combatant.actor.system.details.tier);
-                } else if (combatant.token.disposition < 0) {
-                    enemyShips.push(combatant);
-                    enemyShipTiers.push(combatant.actor.system.details.tier);
-                }
-            }
-        }
-
-        return [playerShips, playerShipTiers, enemyShips, enemyShipTiers];
-    }
-
-    calculatePlayerShipEffectiveTier(shipTiers) {
-        const sortedTiers = shipTiers.sort().reverse(); // sort high to low
-        const highestShipTier = sortedTiers[0];
-        let tierEffective = highestShipTier;
-
-        if (sortedTiers.length > 1) {
-            // count the number of starships within 2 Tiers of the highest one
-            let withinTwo = -1;
-            for (const num of sortedTiers) {
-                if (num >= highestShipTier - 2) {
-                    withinTwo += 1;
-                }
-            }
-
-            // If there are any ships within 2 tiers of the highest one, add one to the effective tier for each of them
-            // Otherwise, sum the remaining ships' tiers and add 1 if the value is equal to or greater than the highest ship's tier
-            if (withinTwo > 0) {
-                tierEffective += withinTwo;
-            } else {
-                const sum = sortedTiers.reduce((partialSum, a) => partialSum + a, 0) - highestShipTier;
-                if (sum >= highestShipTier) {
-                    tierEffective += 1;
-                }
-            }
-        }
-
-        // Calculate other versions of the tier (number, rounded, string)
-        let tierRound = tierEffective;
-        let tierString = "";
-
-        if (tierEffective > 1) {
-            tierRound = Math.floor(tierEffective);
-        }
-        switch (tierRound) {
-            case 0.25:
-                tierString = "1/4";
-                break;
-            case 1 / 3:
-                tierString = "1/3";
-                break;
-            case 0.5:
-                tierString = "1/2";
-                break;
-            default:
-                tierString = String(tierRound);
-        }
-        return [tierEffective, tierRound, tierString];
-    }
-
-    calculateEnemyShipEffectiveTier(shipTiers) {
-        const sortedTiers = shipTiers.sort().reverse(); // sort high to low
-        const highestShipTier = sortedTiers[0];
-        const numShips = shipTiers.length;
-        let tierEffective = highestShipTier;
-
-        if (numShips > 1) {
-            if (sortedTiers[1] === sortedTiers[0]) {
-                tierEffective += 2;
-                if (numShips > 2) {
-                    if (sortedTiers[2] === sortedTiers[0]) {
-                        tierEffective += 1;
-                        if (numShips > 3) {
-                            if (sortedTiers[3] === sortedTiers[0]) {
-                                return this.calculatePlayerShipEffectiveTier(shipTiers);
-                            }
-                        }
-                    }
-                }
-            } else {
-                return this.calculatePlayerShipEffectiveTier(shipTiers);
-            }
-        }
-
-        // Calculate other versions of the tier (number, rounded, string)
-        let tierRound = tierEffective;
-        let tierString = "";
-
-        // Round down if it's not an allowed fractional value (between 0 and 1)
-        if (tierEffective > 1) {
-            tierRound = Math.floor(tierEffective);
-        }
-        switch (tierRound) {
-            case 0.25:
-                tierString = "1/4";
-                break;
-            case 1 / 3:
-                tierString = "1/3";
-                break;
-            case 0.5:
-                tierString = "1/2";
-                break;
-            default:
-                tierString = String(tierRound);
-        }
-        return [tierEffective, tierRound, tierString];
-    }
-
-    /**
-     * Calculate the difficulty of the combat encounter
-     * @returns {[string, number, number]}
-     */
-    calculateShipChallenge() {
-
-        const CRTable = CONFIG.SFRPG.CRTable;
-        const numPlayerShips = this.flags.sfrpg.playerShips.length;
-        const numEnemyShips = this.flags.sfrpg.enemyShips.length;
-        const playerEffectiveTier = this.flags.sfrpg.playerTier.effective;
-        const enemyNumericalTier = this.flags.sfrpg.enemyTier.round;
-        const enemyStringTier = this.flags.sfrpg.enemyTier.string;
-
-        // Check that Players and NPCs are both present
-        if (!numPlayerShips) {
-            console.log('no players');
-            return ["noPCShips", 0, 0];
-        } else if (!numEnemyShips) {
-            console.log('no enemies');
-            return ["noEnemyShips", 0, 0];
-        }
-
-        // Calculate Difficulty Table
-        const diffTable = [
-            {"difficulty": "easy", "tier": playerEffectiveTier - 3},
-            {"difficulty": "average", "tier": playerEffectiveTier - 2},
-            {"difficulty": "challenging", "tier": playerEffectiveTier - 1},
-            {"difficulty": "hard", "tier": playerEffectiveTier},
-            {"difficulty": "epic", "tier": playerEffectiveTier + 1}
-        ];
-
-        // Round down if the tier in the table is a non-integer that isn't allowed
-        for (const row of diffTable) {
-            if (row.tier < 0 || row.tier > 1) {
-                row.tier = Math.floor(row.tier);
-            }
-        }
-
-        // Calculate the Encounter Difficulty
-        let encounterDifficulty = "";
-
-        if (enemyNumericalTier < diffTable[0].tier) {
-            encounterDifficulty = "lessThanEasy";
-        } else if (enemyNumericalTier > diffTable[4].tier) {
-            encounterDifficulty = "greaterThanEpic";
-        } else {
-            for (const diffRow of diffTable) {
-                if (enemyNumericalTier === diffRow.tier) {
-                    encounterDifficulty = diffRow.difficulty;
-                }
-            }
-        }
-
-        return [encounterDifficulty, CRTable[enemyStringTier].totalXP, CRTable[enemyStringTier].wealthValue];
-    }
-
-    /**
-     * Performs all encounter difficulty calculations for normal combat
-     */
-    getNormalEncounterInfo() {
-        if (!this.flags.sfrpg) {
-            this.flags.sfrpg = {};
-        }
-
-        const [PCs, APL] = this.calculateAPL();
-        this.flags.sfrpg.PCs = PCs;
-        this.flags.sfrpg.APL = APL;
-
-        const [enemies, enemyXP] = this.calculateEnemyXP();
-        this.flags.sfrpg.enemies = enemies;
-        this.flags.sfrpg.enemyXP = enemyXP;
-
-        const [CR, XPArray, difficulty, playerXP, wealth] = this.calculateChallenge();
-        this.flags.sfrpg.CR = CR;
-        this.flags.sfrpg.CRXPbounds = [XPArray.minXP, XPArray.totalXP];
-        this.flags.sfrpg.CRXPboundsString = `${XPArray.minXP} – ${XPArray.totalXP}`;
-        this.flags.sfrpg.difficulty = difficulty;
-        this.flags.sfrpg.playerXP = playerXP;
-        this.flags.sfrpg.wealth = wealth;
-
-        this.flags.sfrpg.leftoverCR = this.calculateLeftoverCR();
-    }
-
-    /**
-     * Calculates the APL of the players in a combat
-     * @returns {[Combatant[], number]}
-     */
-    calculateAPL() {
-        const average = (array) => array.reduce((total, value) => total + value, 0) / array.length;
-
-        let playerCombatants = [];
-        let playerLevels = [];
-
-        // Find all player-owned PCs and get their levels
-        for (const combatant of this.combatants) {
-            if (combatant.actor.type === "character") {
-                if (combatant.players.length > 0) {
-                    playerCombatants.push(combatant);
-                    playerLevels.push(combatant.actor.system.details.level.value);
-                }
-            }
-        }
-
-        // Calculate APL
-        if (!playerCombatants.length) {
-            return [playerCombatants, 0];
-        } else if (playerCombatants.length < 4) {
-            return [playerCombatants, Math.round(average(playerLevels)) - 1];
-        } else if (playerCombatants.length > 5) {
-            return [playerCombatants, Math.round(average(playerLevels)) + 1];
-        } else {
-            return [playerCombatants, Math.round(average(playerLevels))];
-        }
-    }
-
-    /**
-     * Calculates the sum of enemy XP values in a combat
-     * @return {[Combatant[], number]}
-    */
-    calculateEnemyXP() {
-        let enemyCombatants = [];
-        let enemyXP = [];
-
-        // Find all player-owned PCs and get their levels
-        for (const combatant of this.combatants) {
-            if (combatant.actor.type === "npc2" || combatant.actor.type === "npc") {
-                if (combatant.isNPC) {
-                    if (combatant.token.disposition < 0) {
-                        enemyCombatants.push(combatant);
-                        enemyXP.push(combatant.actor.system.details.xp.value); // Enemy XP values
-                    }
-                }
-            }
-        }
-        return [enemyCombatants, enemyXP.reduce((partialSum, a) => partialSum + a, 0)];
-    }
-
-    /**
-     * Calculates combat encounter CR, difficulty, and XP value
-     * @returns {[number, object, string, number, number]}
-     */
-    calculateChallenge() {
-        const CRTable = CONFIG.SFRPG.CRTable;
-        const APL = this.flags.sfrpg.APL;
-        const numPlayers = this.flags.sfrpg.PCs.length;
-        const numEnemies = this.flags.sfrpg.enemies.length;
-        const XPtotal = this.flags.sfrpg.enemyXP;
-
-        // Check that Players and NPCs are both present
-        if (!numPlayers) {
-            return ["0", CRTable["0"], "noPcs", 0];
-        } else if (!numEnemies) {
-            return ["0", CRTable["0"], "noEnemies", 0];
-        }
-
-        // Calculate XP and compare to XP table to get CR and wealth value
-        let encounterCR = 0;
-        let XParray = {};
-
-        // Error checking
-        if (XPtotal > CRTable["25"].totalXP) {
-            console.log("Error, encounter CR > 25.");
-            XParray = CRTable["25"];
-            encounterCR = "25";
-        } else {
-            for (let [CR, XProw] of Object.entries(CRTable)) {
-                // Figure out encounter difficulty
-                if (XPtotal <= XProw.totalXP) {
-                    if (XPtotal > XProw.minXP) {
-                        XParray = XProw;
-                        encounterCR = CR;
-                        break;
-                    }
-                }
-            }
-        }
-
-        // Calculate Difficulty Table
-        const diffTable = [
-            {"difficulty": "easy", "CR": APL - 1},
-            {"difficulty": "average", "CR": APL},
-            {"difficulty": "challenging", "CR": APL + 1},
-            {"difficulty": "hard", "CR": APL + 2},
-            {"difficulty": "epic", "CR": APL + 3}
-        ];
-
-        // Calculate a numerical version of the CR for comparison
-        const CRsplit = encounterCR.split("/");
-        let numCR = 0;
-        if (CRsplit.length === 2) {
-            numCR = Number(CRsplit[0]) / Number(CRsplit[1]);
-        } else {
-            numCR = Number(CRsplit[0]);
-        }
-
-        // Calculate the Encounter Difficulty
-        let encounterDifficulty = "";
-
-        if (numCR < diffTable[0].CR) {
-            encounterDifficulty = "lessThanEasy";
-        } else if (numCR > diffTable[4].CR) {
-            encounterDifficulty = "greaterThanEpic";
-        } else {
-            for (const diffRow of diffTable) {
-                if (numCR === diffRow.CR) {
-                    encounterDifficulty = diffRow.difficulty;
-                }
-            }
-        }
-
-        // Calculate individual XP based on number of party members
-        let perPlayerXP = 0;
-
-        if (numPlayers < 4) {
-            perPlayerXP = XParray.perPlayerXP[0];
-        } else if (numPlayers > 5) {
-            perPlayerXP = XParray.perPlayerXP[2];
-        } else {
-            perPlayerXP = XParray.perPlayerXP[1];
-        }
-
-        return [encounterCR, XParray, encounterDifficulty, perPlayerXP, XParray.wealthValue];
-    }
-
-    /**
-     *  Calculate the CR value of the strongest enemy that can be added to the encounter without changing CR
-     * @returns {String}
-     */
-    calculateLeftoverCR() {
-        const enemyXP = this.flags.sfrpg.enemyXP;
-        const remainingXP = this.flags.sfrpg.CRXPbounds[1] - enemyXP;
-        const CRTable = CONFIG.SFRPG.CRTable;
-
-        if (!remainingXP) {
-            return "0";
-        } else if (remainingXP > CRTable["25"].totalXP) {
-            return "25";
-        } else {
-            // Find the CR value of the remaining XP, without going over
-            for (let [CR, XProw] of Object.entries(CRTable)) {
-                if (remainingXP < XProw.nextXP) {
-                    if (remainingXP >= XProw.totalXP) {
-                        return CR;
-                    } else {
-                        return "0";
-                    }
-                }
-            }
-        }
-    }
-
     renderCombatPhase() {
         let phaseDisplay = document.createElement("h4");
         phaseDisplay.classList.add("combat-type");
@@ -972,26 +572,6 @@ export class CombatSFRPG extends Combat {
         combatTypeControls.classList.add("combat-type");
         combatTypeControls.innerHTML = `${prevCombatTypeButton} &nbsp; ${this.getCombatName()} &nbsp; ${nextCombatTypeButton}`;
         document.getElementsByClassName('combat-tracker-header')[0].appendChild(combatTypeControls);
-    }
-
-    renderDifficulty() {
-        const difficulty = this.flags.sfrpg.difficulty;
-        const combatType = this.flags.sfrpg.combatType;
-
-        let difficultyContainer = document.createElement("div");
-        difficultyContainer.classList.add("combat-difficulty-container");
-
-        let difficultyHTML = document.createElement("a");
-        difficultyHTML.classList.add("combat-difficulty", difficulty);
-        if (combatType === 'normal') {
-            difficultyHTML.title = `${game.i18n.format("SFRPG.Combat.Difficulty.Tooltip.PCs")}: ${this.flags.sfrpg.PCs.length} [${game.i18n.format("SFRPG.Combat.Difficulty.Tooltip.APL")} ${this.flags.sfrpg.APL}]\n${game.i18n.format("SFRPG.Combat.Difficulty.Tooltip.HostileNPCs")}: ${this.flags.sfrpg.enemies.length} [${game.i18n.format("SFRPG.Combat.Difficulty.Tooltip.CR")} ${this.flags.sfrpg.CR}]`;
-        } else if (combatType === 'starship') {
-            difficultyHTML.title = game.i18n.format("SFRPG.Combat.Difficulty.Tooltip.ClickForDetails");
-        }
-        difficultyHTML.innerHTML = `Difficulty: ${CONFIG.SFRPG.difficultyLevels[difficulty]}`;
-
-        difficultyContainer.appendChild(difficultyHTML);
-        document.getElementsByClassName('combat-tracker-header')[0].appendChild(difficultyContainer);
     }
 
     _getInitiativeFormula(combatant) {
@@ -1145,7 +725,10 @@ Hooks.on('renderCombatTracker', (app, html, data) => {
 
     const header = html.find('.combat-tracker-header');
     const footer = html.find('.directory-footer');
+
+    // Return whether the difficulty tracker should be displayed
     const diffDisplay = game.settings.get("sfrpg", "difficultyDisplay");
+    const diffObject = new CombatDifficulty(activeCombat);
 
     if (activeCombat.round) {
         const phases = activeCombat.getPhases();
@@ -1156,16 +739,16 @@ Hooks.on('renderCombatTracker', (app, html, data) => {
         // Add buttons for switching combat type
         activeCombat.renderCombatTypeControls();
 
-        // Add in the difficulty calculator if needed
+        // Add in the normal encounter difficulty calculator if needed
         if (activeCombat.getCombatType() === "normal" && game.user.isGM && diffDisplay) {
-            activeCombat.getNormalEncounterInfo();
-            activeCombat.renderDifficulty();
+            diffObject.getNormalEncounterInfo();
+            diffObject.renderDifficulty();
         }
 
-        // Add in the difficulty calculator if needed
+        // Add in the starship encounter difficulty calculator if needed
         if (activeCombat.getCombatType() === "starship" && game.user.isGM && diffDisplay) {
-            activeCombat.getStarshipEncounterInfo();
-            activeCombat.renderDifficulty();
+            diffObject.getStarshipEncounterInfo();
+            diffObject.renderDifficulty();
         }
 
         // Handle button clicks
@@ -1192,11 +775,11 @@ Hooks.on('renderCombatTracker', (app, html, data) => {
             ev.preventDefault();
             console.log("Starfinder | Rendering Encounter Statistics Dialog");
             const isStarship = activeCombat.flags.sfrpg.combatType === "starship";
-            const contentTemplate = `//systems/sfrpg/templates/combat/${isStarship ? "starship" : "normal"}-encounter-stats.hbs`;
+            const contentTemplate = `//systems/sfrpg/templates/apps/${isStarship ? "starship" : "normal"}-encounter-stats.hbs`;
 
             new Dialog({
-                title: `${game.i18n.format("SFRPG.Combat.Difficulty.Tooltip.Details")}: ${CONFIG.SFRPG.difficultyLevels[activeCombat.flags.sfrpg.difficulty]} ${game.i18n.format("SFRPG.Combat.Difficulty.Tooltip.DifficultyEncounter")}`,
-                content: await renderTemplate(contentTemplate, activeCombat),
+                title: `${game.i18n.format("SFRPG.Combat.Difficulty.Tooltip.Details")}: ${CONFIG.SFRPG.difficultyLevels[diffObject.difficultyData.difficulty]} ${game.i18n.format("SFRPG.Combat.Difficulty.Tooltip.DifficultyEncounter")}`,
+                content: await renderTemplate(contentTemplate, diffObject),
                 buttons: {},
                 resizable: true
             }, {id: "encounter-stats"}).render(true);
